Add CommonJWTValidatorImpl
