+++ components/apimgt/org.wso2.carbon.apimgt.common.gateway/src/main/java/org/wso2/carbon/apimgt/common/gateway/bootstrap/Bootstrap.java
@@ -1,0 +1,56 @@
+/*
+ *  Copyright (c) 2022, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.wso2.carbon.apimgt.common.gateway.bootstrap;
+
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.wso2.carbon.apimgt.common.gateway.configdto.HttpClientConfigurationDTO;
+import org.wso2.carbon.apimgt.common.gateway.util.CommonAPIUtil;
+
+/**
+ * Bootstrap contains the common gateway package-wise configurations and instances.
+ */
+public class Bootstrap {
+
+    private static final Bootstrap bootstrap = new Bootstrap();
+    private volatile CloseableHttpClient httpClient;
+    private HttpClientConfigurationDTO httpClientConfiguration;
+
+    private Bootstrap() {
+    }
+
+    public static Bootstrap getInstance() {
+        return bootstrap;
+    }
+
+    public void setHttpClientConfiguration(HttpClientConfigurationDTO httpClientConfigurationDTO) {
+        httpClientConfiguration = httpClientConfigurationDTO;
+    }
+
+    public CloseableHttpClient getHttpClient() {
+        if (httpClient == null) {
+            synchronized (new Object()) {
+                if (httpClient == null) {
+                    // protocol is unused
+                    httpClient = (CloseableHttpClient)
+                            CommonAPIUtil.getHttpClient("http", httpClientConfiguration);
+                }
+            }
+        }
+        return httpClient;
+    }
+}
